feature(*): exports separate editors
